Code: break HomePage into components
